Clean up ProjectN parts of the compiler
These files were exclusively used for .NET Native to support the UTC codegen backend.

I left two things that we don't use in CoreRT but might want to use:

1. Support for SharedLibrary
2. Windows-specific managed debugger support

I'm not very firm on keeping 2.
