+++ src/ILCompiler.Compiler/src/Compiler/DependencyAnalysis/GenericLookupResult.cs
@@ -728,9 +728,6 @@ namespace ILCompiler.DependencyAnalysis
         public override void WriteDictionaryTocData(NodeFactory factory, IGenericLookupResultTocWriter writer)
         {
             LookupResultType lookupResult = LookupResultType.Method;
-            if (_isUnboxingThunk && ProjectNDependencyBehavior.EnableFullAnalysis)
-                lookupResult = LookupResultType.UnboxingMethod;
-
             writer.WriteData(LookupResultReferenceType(factory), lookupResult, _method);
         }
 
@@ -1256,10 +1253,7 @@ namespace ILCompiler.DependencyAnalysis
 
         public override ISymbolNode GetTarget(NodeFactory factory, GenericLookupResultContext dictionary)
         {
-            UtcNodeFactory utcNodeFactory = factory as UtcNodeFactory;
-            Debug.Assert(utcNodeFactory != null);
-            TypeDesc instantiatedType = _type.GetNonRuntimeDeterminedTypeFromRuntimeDeterminedSubtypeViaSubstitution(dictionary.TypeInstantiation, dictionary.MethodInstantiation);
-            return utcNodeFactory.TypeThreadStaticsIndexSymbol((MetadataType)instantiatedType);
+            throw new NotImplementedException();
         }
 
         public override void AppendMangledName(NameMangler nameMangler, Utf8StringBuilder sb)
@@ -1315,11 +1309,7 @@ namespace ILCompiler.DependencyAnalysis
 
         public override ISymbolNode GetTarget(NodeFactory factory, GenericLookupResultContext dictionary)
         {
-            UtcNodeFactory utcNodeFactory = factory as UtcNodeFactory;
-            Debug.Assert(utcNodeFactory != null);
-            TypeDesc instantiatedType = _type.GetNonRuntimeDeterminedTypeFromRuntimeDeterminedSubtypeViaSubstitution(dictionary.TypeInstantiation, dictionary.MethodInstantiation);
-            Debug.Assert(instantiatedType is MetadataType);
-            return utcNodeFactory.TypeThreadStaticsOffsetSymbol((MetadataType)instantiatedType);
+            throw new NotImplementedException();
         }
 
         public override void AppendMangledName(NameMangler nameMangler, Utf8StringBuilder sb)
