+++ src/ILCompiler.Compiler/src/Compiler/DependencyAnalysis/MrtImportImportedNodeProvider.cs
@@ -10,6 +10,19 @@ using Debug = System.Diagnostics.Debug;
 
 namespace ILCompiler.DependencyAnalysis
 {
+    public class ImportExportOrdinals
+    {
+        public bool isImport;
+        public uint tlsIndexOrdinal;
+        public ReadOnlyDictionary<TypeDesc, uint> typeOrdinals;
+        public ReadOnlyDictionary<TypeDesc, uint> nonGcStaticOrdinals;
+        public ReadOnlyDictionary<TypeDesc, uint> gcStaticOrdinals;
+        public ReadOnlyDictionary<TypeDesc, uint> tlsStaticOrdinals;
+        public ReadOnlyDictionary<MethodDesc, uint> methodOrdinals;
+        public ReadOnlyDictionary<MethodDesc, uint> unboxingStubMethodOrdinals;
+        public ReadOnlyDictionary<MethodDesc, uint> methodDictionaryOrdinals;
+    }
+
     public class MrtImportImportedNodeProvider : ImportedNodeProvider
     {
         private readonly KeyValuePair<MrtProcessedImportAddressTableNode, ImportExportOrdinals>[] _importOrdinals;
